Implemented store abstraction
Implemented store abstraction to allow the implementation of a persistent data base later down the line.

Also resolved a small formatting error in a HTTP response.

Skipping the JWT authentication for now as it was causing a lot of unexpected errors. Need to investigate the subject a bit more and try again later down the line.